pyproject.toml
[project]
name = "fambot-backend"
version = "0.1.0"
description = "FastAPI backend: JWT auth, Firestore profiles, cardiovascular risk scoring (cardio_train pipeline)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "xgboost>=3.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "firebase-admin>=6.5.0",
    "pydantic>=2.10.0",
    "email-validator>=2.0.0",
    "joblib>=1.4.0",
    "PyJWT>=2.9.0",
    "python-dotenv>=1.2.2",
    "segno>=1.6.1",
    "python-multipart>=0.0.20",
    "google-genai>=1.13.0",
]

[project.scripts]
model = "model:main"
api = "fambot_backend.app:run"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["fambot_backend*"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
]

[tool.setuptools]
py-modules = ["model"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "unit: pure logic, no network or cloud I/O",
    "api: HTTP layer via TestClient; mocks or FAMBOT_SKIP_* flags",
    "external: real network or Google APIs; requires FAMBOT_RUN_EXTERNAL_TESTS=1",
]
addopts = "-m 'not external'"
